Initial Version
Modern watchface (thanks to Zalew) with the ability to pick up calendar
events from Smartwatch Pro (thanks to Max Baeumle) and create alerts
from them.

After starting and every 10 minutes a calendar request is issued. Any
event which is for today, has a time and alerts has alerts generated
for it.

When an alert triggers the usual date and day is replaced by the title
of the event with a relative time (e.g. In 15 mins). The vibration and
light are triggered. After 15 seconds the vibration and light are
triggered again. After 30 the display reverts to showing the date.

There are up to three icons under the pebble logo. A right arrow
indicates a request has been made of the iphone. A left arrow indicates
a reply has been recieved. The bell indicates an alert has been created.